Fix warnings and duplicated variables
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@4155 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/COpenGLDriver.h
@@ -502,7 +502,7 @@ namespace video
 
 			const ITexture* operator[](int stage) const
 			{
-				if (stage<MATERIAL_MAX_TEXTURES)
+				if ((u32)stage<MATERIAL_MAX_TEXTURES)
 					return CurrentTexture[stage];
 				else
 					return 0;
